requirements.txt
pyspark==3.1.2
delta-spark==2.3.0
python-dotenv
py4j


